nuevo endpoint busqueda avanzada con mejoras en las firmas de los metodos, tambien describo las peliculas como se eliminan con el confimr

+++ app/api/v1/endpoints/movies.py
@@ -7,20 +7,22 @@ from sqlalchemy.orm import Session
 from fastapi import Depends
 
 
-from app.config.config import config
-import logging
-
-
-logger = logging.getLogger(__name__)
-
 router = APIRouter()
 
-@router.get("/hello")
+@router.get("/hello"
+            , status_code=status.HTTP_200_OK
+            , description="Endpoint de prueba"
+            )
 def read_hello():
     """Endpoint principal de la API."""
     return {"message": "Bienvenido al Catálogo de Películas 🎬"}
 
-@router.post("/", response_model=ApiResponse[MovieResponse])
+
+@router.post("/"
+             , response_model=ApiResponse[MovieResponse]
+             , status_code=status.HTTP_201_CREATED
+             , description="Crea una nueva película"
+             )
 def create_movie(
     request: MovieCreate,
     db: Session = Depends(db_connection.get_db),
@@ -35,7 +37,12 @@ def create_movie(
         data=MovieResponse.model_validate(movie)
     )
 
-@router.get("/", response_model=ApiResponse[list[MovieResponse]])
+
+@router.get("/"
+            , response_model=ApiResponse[list[MovieResponse]]
+            , status_code=status.HTTP_200_OK
+            , description="Listado de todas las películas"
+            )
 def get_movies(
     db: Session = Depends(db_connection.get_db),
 ):
@@ -48,7 +55,36 @@ def get_movies(
         data=[MovieResponse.model_validate(m) for m in movies]
     )
 
-@router.get("/{movie_id}", response_model=ApiResponse[MovieResponse])
+
+@router.get("/buscar"
+            , response_model=ApiResponse[list[MovieResponse]]
+            , status_code=status.HTTP_200_OK
+            , description="Búsqueda por titulo, director o genero total o parcial")
+def search_movies(
+    search: str | None = None,
+    offset: int = 0,
+    fetch: int = 100,
+    db: Session = Depends(db_connection.get_db),
+):
+    repo = MovieRepository(db)
+    movies = repo.search(
+        search=search,
+        offset=offset,
+        fetch=fetch
+        )
+    return ApiResponse(
+        status="success",
+        message="Listado obtenido correctamente",
+        errors=[],
+        data=[MovieResponse.model_validate(m) for m in movies]
+    )
+
+
+@router.get("/{movie_id}"
+            , response_model=ApiResponse[MovieResponse]
+            , status_code=status.HTTP_200_OK
+            , description="Búsqueda por id"
+            )
 def get_movie_by_id(
     movie_id: int,
     db: Session = Depends(db_connection.get_db),
@@ -65,7 +101,12 @@ def get_movie_by_id(
         data=MovieResponse.model_validate(movie)
     )
 
-@router.patch("/{movie_id}", response_model=ApiResponse[MovieResponse])
+
+@router.patch("/{movie_id}"
+              , response_model=ApiResponse[MovieResponse]
+              , status_code=status.HTTP_200_OK
+              , description="Actualiza una pelicula"
+              )
 def update_movie_by_id(
     movie_id: int,
     request: MovieUpdate,
@@ -86,18 +127,31 @@ def update_movie_by_id(
         data=MovieResponse.model_validate(repo.update(movie))
     )
 
+
 @router.delete(
         "/{movie_id}"
         , response_model=ApiResponse[DeleteMovieResponse]
         , status_code=status.HTTP_200_OK
-)
+        , description="Elimina una pelicula por id"
+        )
 def delete_movie_by_id(
     movie_id: int,
+    confirm: bool = True,
     db: Session = Depends(db_connection.get_db),
-):
-    
+) -> ApiResponse[DeleteMovieResponse]:
+    """
+    Elimina una pelicula, permite flag de confirmación util para dry-run
+
+    Args:
+        movie_id (int): id de la pelicula
+        confirm (bool, optional): true si se desea realizar el borrado. Defaults to True.
+        db (Session, optional): session de la base de datos. Defaults to Depends(db_connection.get_db).
+
+    Returns:
+        _type_: ApiResponse
+    """
     repo = MovieRepository(db)    
-    repo.delete_by_id(movie_id)
+    repo.delete_by_id(movie_id, confirm)
 
     return ApiResponse(
         status="success",
